requirements.txt
flask
flask-cors
numpy
pillow
tensorflow-cpu==2.13.0